rename serializer -> persistence
