+++ examples/10_programmatic_load.py
@@ -8,7 +8,7 @@ import os
 import xnmt.tee
 import xnmt.serialize.imports
 from xnmt.param_collection import ParamManager
-from xnmt.serializer import YamlSerializer, YamlPreloader, LoadSerialized
+from xnmt.persistence import YamlSerializer, YamlPreloader, LoadSerialized
 
 EXP_DIR = os.path.dirname(__file__)
 EXP = "programmatic-load"
