+++ xnmt/attender.py
@@ -6,7 +6,7 @@ import dynet as dy
 
 from xnmt.param_collection import ParamManager
 from xnmt.param_init import GlorotInitializer, ZeroInitializer
-from xnmt.serializer import serializable_init, Serializable, Ref, bare
+from xnmt.persistence import serializable_init, Serializable, Ref, bare
 
 class Attender(object):
   '''
