+++ xnmt/bridge.py
@@ -3,7 +3,7 @@ import dynet as dy
 import xnmt.linear
 from xnmt.param_collection import ParamManager
 from xnmt.param_init import GlorotInitializer, ZeroInitializer
-from xnmt.serializer import serializable_init, Serializable, Ref, bare
+from xnmt.persistence import serializable_init, Serializable, Ref, bare
 
 class Bridge(object):
   """
