+++ xnmt/ff.py
@@ -1,7 +1,7 @@
 import dynet as dy
 
 from xnmt.transducer import SeqTransducer, FinalTransducerState
-from xnmt.serializer import Serializable
+from xnmt.persistence import Serializable
 from xnmt.expression_sequence import ExpressionSequence
 from xnmt.param_collection import ParamManager
