+++ xnmt/lstm.py
@@ -8,7 +8,7 @@ from xnmt.events import register_xnmt_handler, handle_xnmt_event
 from xnmt.param_collection import ParamManager
 from xnmt.param_init import GlorotInitializer, ZeroInitializer
 from xnmt.transducer import SeqTransducer, FinalTransducerState
-from xnmt.serializer import serializable_init, Serializable, Ref, bare
+from xnmt.persistence import serializable_init, Serializable, Ref, bare
 
 class UniLSTMState(object):
   """
