+++ xnmt/preproc.py
@@ -14,7 +14,7 @@ with warnings.catch_warnings():
   import h5py
 import yaml
 
-from xnmt.serializer import serializable_init, Serializable
+from xnmt.persistence import serializable_init, Serializable
 from xnmt.speech_features import logfbank, calculate_delta, get_mean_std, normalize
 from xnmt.util import make_parent_dir
