+++ xnmt/preproc_runner.py
@@ -4,7 +4,7 @@ import os.path
 from typing import List
 
 from xnmt.preproc import Normalizer, SentenceFilterer, VocabFilterer
-from xnmt.serializer import serializable_init, Serializable
+from xnmt.persistence import serializable_init, Serializable
 from xnmt.util import make_parent_dir
 
 class PreprocTask(object):
