+++ xnmt/retriever.py
@@ -9,7 +9,7 @@ from simple_settings import settings
 import xnmt.batcher
 from xnmt.events import handle_xnmt_event
 from xnmt.generator import GeneratorModel
-from xnmt.serializer import serializable_init, Serializable
+from xnmt.persistence import serializable_init, Serializable
 from xnmt.reports import Reportable
 from xnmt.expression_sequence import ExpressionSequence
