+++ xnmt/segmenting_encoder.py
@@ -14,7 +14,7 @@ import xnmt.expression_sequence as expression_sequence
 from xnmt.batcher import Mask
 from xnmt.events import register_xnmt_handler, handle_xnmt_event
 from xnmt.reports import Reportable
-from xnmt.serializer import serializable_init, Serializable
+from xnmt.persistence import serializable_init, Serializable
 from xnmt.transducer import SeqTransducer
 from xnmt.loss import LossBuilder
 from xnmt.param_collection import ParamManager
