+++ xnmt/training_task.py
@@ -11,7 +11,7 @@ from xnmt.loss import LossBuilder
 from xnmt.loss_calculator import LossCalculator, MLELoss
 from xnmt.loss_tracker import BatchLossTracker
 from xnmt.param_collection import ParamManager
-from xnmt.serializer import serializable_init, Serializable, bare
+from xnmt.persistence import serializable_init, Serializable, bare
 
 class TrainingTask(object):
   """
