+++ xnmt/transformer.py
@@ -1,7 +1,7 @@
 import numpy as np
 import dynet as dy
 from xnmt.linear import Linear
-from xnmt.serializer import serializable_init, Serializable, Ref
+from xnmt.persistence import serializable_init, Serializable, Ref
 from xnmt.events import register_xnmt_handler, handle_xnmt_event
 from xnmt.param_init import LeCunUniformInitializer
 from xnmt.param_collection import ParamManager
